CicloUno/pandas/requirements.txt
pandas==1.1.5
numpy==1.19.5
python-dotenv==0.17.1
